fix: rename Verdaccio storage folder to the standard `node-modules` name
The TypeScript project configs in the root workspace of this repository were not excluding the files in storage because the folder name `storage` is not a default excluded item.

This led to TypeScript processing thousands of files implicitly included in the project that should have been ignored.

See more at https://github.com/typescript-eslint/typescript-eslint/issues/1192#issuecomment-565775211